src/requirements.txt
django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0